divided app into modules
