+++ package.json
@@ -3,8 +3,9 @@
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "start": "node ./src/index.js",
+    "start": "node ./src/app/index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
